+++ tom_d4rt_generator/example/dart_overview/lib/src/d4rt_bridges/dart_overview_bridges.b.dart
@@ -1,6 +1,6 @@
 // D4rt Bridge - Generated file, do not edit
 // Sources: 16 files
-// Generated: 2026-02-09T05:32:33.062038
+// Generated: 2026-02-09T10:29:23.188123
 
 // ignore_for_file: unused_import, deprecated_member_use, prefer_function_declarations_over_variables