+++ tom_d4rt_generator/example/example_project/lib/dartscript.b.dart
@@ -1,6 +1,6 @@
 // D4rt Bridge - Generated file, do not edit
 // Dartscript registration for d4rt_generator_example
-// Generated: 2026-02-09T05:34:22.628169
+// Generated: 2026-02-09T10:27:59.053175
 
 /// D4rt Bridge Registration for d4rt_generator_example
 library;