+++ tom_d4rt_generator/example/user_reference/bin/d4rtrun.b.dart
@@ -1,6 +1,6 @@
 // D4rt Bridge - Generated file, do not edit
 // Test runner for user_reference_example
-// Generated: 2026-02-09T05:31:37.120686
+// Generated: 2026-02-09T10:28:16.184370
 //
 // Usage:
 //   dart run bin/d4rtrun.b.dart <script.dart|.d4rt>  Run a D4rt script file