+++ tom_d4rt_generator/example/userbridge_override/bin/d4rtrun.b.dart
@@ -1,6 +1,6 @@
 // D4rt Bridge - Generated file, do not edit
 // Test runner for userbridge_override_example
-// Generated: 2026-02-09T05:32:10.323427
+// Generated: 2026-02-09T10:28:56.961606
 //
 // Usage:
 //   dart run bin/d4rtrun.b.dart <script.dart|.d4rt>  Run a D4rt script file